+++ src/main.cpp
@@ -1,89 +1,61 @@
-#include <iostream>
 #include <vector>
+#include <chrono>
+#include <csignal>
+#include <iostream>
+#include <memory>
+#include <sstream>
+#include <thread>
 
 #include "lib/option.hpp"
 #include "lib/result.hpp"
+#include "lib/server.hpp"
 
 using namespace std;
 
-/// Funcion que intenta pedir un numero al usuario en un rango.
-/// En caso de ingresar un numero que no está en rango, se devolverá un mensaje de error.
-/// En caso de exito, se devuelve el valor ingresado.
-Result<int, string> try_range(int from, int to) {
-  int result;
-  cin >> result;
+volatile sig_atomic_t st = 0;
 
-  if (result >= to || result < from) {
-    return Result<int, string>::Err("Error: El valor no está en el rango especificado");
-  } else {
-    return Result<int, string>::Ok(result);
-  }
+void inthand(int signum) {
+  kill_sign();
+  st = 1;
 }
 
+void resolve(shared_ptr<Connection> s, TcpServer &tcp) {
+  if (s->is_null())
+    return;
+  stringstream a((string()));
+  string r = s->get_msg();
 
+  a << "Hola desde C++ " << r << "!" << endl;
 
-template<typename K, typename V>
-struct KeyVal {
-  K key;
-  V value;
-};
-
-template<typename K, typename V>
-struct Dict {
-  vector<KeyVal<K, V>> items;
-};
-
-struct Index {
-  vector<KeyVal<string, string>> index_table;
-
-  Index() : index_table(vector<KeyVal<string, string>>()) {}
-};
-
-struct Table {
-  Index index;
-  vector<KeyVal<string, size_t>> signature;
-  Table() : index(Index()) {}
-};
-
-class DB {
-protected:
-  vector<Table> tables;
-public:
-  DB() : tables(vector<Table>()) {};
-
-  void create_table(const string & name, const string & pk) {
-
+  if (r == "slow") {
+    this_thread::sleep_for(chrono::seconds(15));
   }
-};
+
+  if (r == "stop")
+    inthand(0);
+  else
+    tcp.send(*s, a.str());
+}
 
 int main() {
-  DB db = DB();
+  cout << "Starting CppServer 0.1.11 ..." << endl;
+  vector<thread> threads;
 
-  // Option: Al igual que en rust puede ser un Some o un None.
-  // Nos sirve para respresentar la ausencia de un dato de forma explicita,
-  // así podemos hacer un mejor manejo de ausencia de datos.
-  Option<int> v = Option<int>::Some(4);
+  signal(SIGINT, inthand);
+  signal(SIGTERM, inthand);
 
-  // Lo siguiente crashearía porque no hubo un chequeo de status con v.is_some() o v.is_none()
-  // cout << "El valor es " << v.inner() << endl;
+  {
+    TcpServer server = TcpServer();
 
-  if (v.is_some()) {
-    cout << "El valor es " << v.inner() << endl;
-  } else {
-    cout << "No hay valor" << endl;
+    while (!st) {
+      shared_ptr<Connection> result = server.recv();
+      threads.emplace_back(
+          thread([result, &server] { resolve(result, server); }));
+    }
   }
 
-  // Result es una versión mejorada de Opcion que permite representar la ausencia de estado junto con
-  // un dato que indique de alguna manera qué falló.
-  // En este caso la funcion try_range puede devolver un int en caso de exito o un
-  // string con un mensaje que comunica la causa del error.
-  Result<int, string> resultado = try_range(10, 20);
-
-  if (resultado.is_ok()) {
-    cout << "Ingresaste el valor: " << resultado.inner() << endl;
-  } else
-    cout << resultado.inner_err() << endl;
-
-  return 0;
+  cout << "Finishing CPP..." << endl;
+  for (auto &t : threads) {
+    t.join();
+  }
 }
-
