--- tcp_server_experiment/justfile
@@ -1,26 +1,0 @@
-default:
-  just --list
-
-cargo_build_release:
-  cd server && cargo build --release
-
-cargo_build:
-  cd server && cargo build
-
-init:
-  cmake .
-
-build: init cargo_build
-  cmake --build .
-
-build_release: init cargo_build_release
-  cmake --build . --config Release
-
-run: build
-  ./tcp_server_experiment
-
-run_release: build_release
-  ./tcp_server_experiment
-
-clean:
-  rm -rf cmake_install.cmake CMakeFiles CMakeCache.txt Makefile
